__nextcloud: stricter check configuration values
This now checks the configuration value exactly and does not say there
are equal if just the end of the string differs.

Resulting from this, it will run code the next run if the default port
was in a database migration. This was fixed by instantly correct this
configuration value.