[__jitsi_meet] Update to 2.3, add versioning parameter.
Jitsi's repositories have rotated keys, this removes the previous key if
necessary and ensures the new key is present.

Can't be merged until following lands:
  https://code.ungleich.ch/ungleich-public/cdist/-/merge_requests/994
since we rely on `--use-deprecated-apt-key` and the improvements in the type to
modify the keyring in a reliable fashion.

This also updates the exporter to version 1.1.5 released on April 25th 2021.