__nextcloud upgrading safety check (#3)
* __nextcloud: finer version security check

This implements a version check that you can't skip major Nextcloud
versions except explicitly wanted. Also, the manual was updated to
explain more and more about the corner cases the type can not check
(sure, but would be more difficult cause you need to request external
information).

* __nextcloud: fix major version comparsion

The awk implementation was not correctly changed from the above usage in
the version_ge function. Was corrected now.

* __nextcloud: add a few comments

Few comments, but mainly cause I've tried to update all apps before
an upgrade, which looks like it is not necessary to do so.

* __nextcloud: fixed review remarks

AWK check was shortend and explicit return code statement was removed
because it will be done implicit.