__netbox: pass handling of requirements.txt to pip3
Previous handling passed a list of pip packages from the
requirements.txt via xargs to the pip install directly.

This is error-prone, as shown with the major 3 Netbox release.
This type breaks cause of comments inside of it. This commit
fixes it, while keeping the compatibility to install gnuicorn
separate.